Added RHS vector output to a file
